Projects/llm_image_editor/pyproject.toml
[tool.poetry]
name = "llm-image-editor"
version = "0.1.0"
description = ""
authors = ["JHyuk2 <whdgurdl77@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.41.1"
opencv-python = "^4.10.0.84"
pillow = "^11.0.0"
ipykernel = "^6.29.5"
diffusers = "^0.31.0"
transformers = "^4.47.0"
accelerate = "^1.2.1"
torch = "^2.5.1"
torchvision = "^0.20.1"
streamlit-drawable-canvas = "^0.9.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
